requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-groq==0.0.1

sentence-transformers==2.2.2
streamlit==1.29.0
pymupdf==1.23.14
python-dotenv==1.0.0
tiktoken==0.5.2

# Google Gemini support
google-generativeai>=0.3.0

# HTTP requests for Ollama and API validation
requests>=2.28.0
